--- kitchen_feud/Assets/Scripts/AddUIElements.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c49ab0921a6943149b99a1ca5821114a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
