+++ kitchen_feud/Assets/Scripts/Canvas/gameOverMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7cf9fa5921dae5d43bba415a78498a6b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
