--- kitchen_feud/Assets/Scripts/HealthBar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a16536f4b176a4b38af63215e403eeeb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
