+++ kitchen_feud/Assets/Scripts/Items/TrayController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5afc471bd737cb446a51232edb36121d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
