--- kitchen_feud/Assets/Scripts/Pantry.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5dc47c2d3a9015640b74b7038667e607
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
