+++ kitchen_feud/Assets/Scripts/Particle Effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fafb38783f2523245aef8f702fd25d09
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
