+++ kitchen_feud/Assets/Scripts/Particle Effects/ProjectileLifetime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d9eb73b90180e3444bd2068f49ba8eb9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
