--- kitchen_feud/Assets/Scripts/PickUpController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 187310dc8aba213a8a3760a35a7e3ca6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
