+++ kitchen_feud/Assets/Scripts/Player/HealthBar.cs
@@ -1,0 +1,28 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class HealthBar : MonoBehaviour
+{
+
+	public Slider slider;
+	public Gradient gradient;
+	public Image fill;
+
+	public void SetMaxHealth(float health)
+	{
+		slider.maxValue = health;
+		slider.value = health;
+
+		fill.color = gradient.Evaluate(1f);
+	}
+
+    public void SetHealth(float health)
+	{
+		slider.value = health;
+
+		fill.color = gradient.Evaluate(slider.normalizedValue);
+	}
+
+}
