+++ kitchen_feud/Assets/Scripts/Player/MovingInstructions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8cb706131e7048eda73ce187419bb05
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
