+++ kitchen_feud/Assets/Scripts/Player/kickPlayers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 617e077bac08d4f0aad47c717d3b2460
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
