--- kitchen_feud/Assets/Scripts/SmokeGrenade.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b9f39460769f66408d3184619f49bb8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
