--- kitchen_feud/Assets/Scripts/SpawnPlayers.cs
@@ -1,24 +1,0 @@
-using UnityEngine;
-using Photon.Pun;
-
-public class SpawnPlayers : MonoBehaviour
-{
-    public GameObject playerPrefab;
-
-    public float minX;
-    public float minY;
-    public float minZ;
-    public float maxX;
-    public float maxY;
-    public float maxZ;
-    public static GameObject LocalPlayerInstance;
-
-
-    private void Start()
-    {
-        //PhotonNetwork.IsMessageQueueRunning = true;
-        Vector3 randomPosition = new Vector3(Random.Range(minX, maxX), Random.Range(minY, maxY), Random.Range(minZ, maxZ));
-        PhotonNetwork.Instantiate(playerPrefab.name, randomPosition, Quaternion.identity);
-        //DontDestroyOnLoad(this.gameObject);
-    }
-}
