--- kitchen_feud/Assets/Scripts/SpawnPlayers.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 80255c67b7a794319b06b52631cbb93f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
