--- kitchen_feud/Assets/Scripts/smoke.cs
@@ -1,54 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class smoke : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-
-    void smokeEffect(){
-		// particleSystem = appliance.GetComponent<ParticleSystem>();
-		// var main = particleSystem.main;
-		// var emission = particleSystem.emission;
-
-		// bool isStopped = particleSystem.isStopped;
-		// float duration = 20 + (10 * score/slider.maxValue);
-
-		// if (isStopped){
-		// 	main.duration = duration;
-		// }else if (prevScore != score){
-		// 	particleSystem.Stop();
-		// 	particleSystem.Clear();
-		// 	main.duration = (main.duration + duration)/2 ;
-		// }
-		
-		// appliance.GetComponentInChildren<ParticleSystem>().Play();
-
-		// float startSpeed = 0.5f + (0.25f * score/slider.maxValue);
-		// float rateOverTime;
-		// if (score > 0){
-		// 	rateOverTime = 15 + (25 * score/slider.maxValue);
-		// }else{
-		// 	rateOverTime = 15;
-		// }
-
-
-		// if (!isStopped){
-		// 	main.startSpeed = (main.startSpeed.constant + startSpeed)/2;
-		// 	emission.rateOverTime = (emission.rateOverTime.constant + rateOverTime)/2;
-		// }else{
-		// 	main.startSpeed = startSpeed;
-		// 	emission.rateOverTime = rateOverTime;
-		}
-}
