--- kitchen_feud/Assets/Scripts/smoke.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9ddb0eb96470c9f4abef880b20e018f7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
